# security/openapi-ci.yaml
openapi: 3.0.3
info:
  title: BlockGo Middleware Security-Test API
  version: 2.0.0
  description: |
    CI contract for the public Node middleware gateway. It documents the routes
    owned by auth, Fabric identity, ledger, grade upload, and settings services.
    The server is intentionally fixed to the isolated GitHub Actions runner.
servers:
  - url: http://127.0.0.1:4000
...
paths:
...
  /api/reset-password:
    post:
      tags: [Authentication]
      operationId: resetPassword
      summary: Complete password reset with an OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, otp, newPassword]
              additionalProperties: false
              properties:
                email: { type: string, format: email, maxLength: 255 }
                otp: { type: string, pattern: "^[0-9]{6}$" }
                newPassword: { type: string, minLength: 8, maxLength: 128 }
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: The OTP is invalid or expired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Error:
      type: object
      required: [error]
      additionalProperties: true
      properties:
        error: { type: string }
...
  responses:
...
    Message:
      description: Generic message response
      content:
        application/json:
          schema:
            type: object
            required: [message]
            properties:
              message: { type: string }
...
    BadRequest:
      description: Invalid or malformed request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InternalError:
      description: Internal service failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }